+++ src/unversioned/get_supported_versions.rs
@@ -1,4 +1,6 @@
-//! [GET /_matrix/client/versions](https://matrix.org/docs/spec/client_server/r0.4.0.html#get-matrix-client-versions)
+//! [GET /_matrix/client/versions](https://matrix.org/docs/spec/client_server/r0.6.0.html#get-matrix-client-versions)
+
+use std::collections::HashMap;
 
 use ruma_api::ruma_api;
 
@@ -17,5 +19,8 @@ ruma_api! {
     response {
         /// A list of Matrix client API protocol versions supported by the homeserver.
         pub versions: Vec<String>,
+        /// Experimental features supported by the server.
+        #[serde(default, skip_serializing_if = "HashMap::is_empty")]
+        pub unstable_features: HashMap<String, bool>
     }
 }