+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Improvements:
 * Add `r0::client_exchange::send_event_to_device` (introduced in r0.3.0)
 * Add endpoints to retrieve account_data (introduced in r0.5.0)
 * Add media endpoints: `r0::media::{get_media_config, get_media_preview, get_content_as_filename}`
+* Add `unstable_features` to `unversioned::get_supported_versions` (introduced in r0.5.0)
 
 Breaking changes: